requirements.txt
testtools>=2.4.0
docker>=4.2.0
tzlocal>=2.1
requests
#linting
flake8
black==22.8.0
mypy
types-pycurl
types-requests
types-PyYAML
types-urllib3
types-docker
#
pandas>=1.1.3
numpy
scipy
#
Pillow
pycurl>=7.43
click>=7.1.2
rich
patch-ng

interrogate==1.7.0
